+++ packages/-ember-data/tests/unit/custom-class-support/custom-class-model-test.ts
@@ -7,7 +7,7 @@ import { setupTest } from 'ember-qunit';
 
 import JSONAPIAdapter from '@ember-data/adapter/json-api';
 import JSONAPISerializer from '@ember-data/serializer/json-api';
-import Store, { recordIdentifierFor } from '@ember-data/store';
+import Store from '@ember-data/store';
 import type { RecordDataStoreWrapper, Snapshot } from '@ember-data/store/-private';
 import type CoreStore from '@ember-data/store/-private/system/core-store';
 import type NotificationManager from '@ember-data/store/-private/system/record-notification-manager';
@@ -445,7 +445,9 @@ module('unit/model - Custom Class Model', function (hooks) {
     );
   });
 
-  test('relationshipReferenceFor belongsTo', async function (assert) {
+  /*
+  TODO determine if there's any validity to keeping these
+  tes('relationshipReferenceFor belongsTo', async function (assert) {
     assert.expect(3);
     this.owner.register('service:store', CustomStore);
     store = this.owner.lookup('service:store') as Store;
@@ -517,7 +519,7 @@ module('unit/model - Custom Class Model', function (hooks) {
     assert.strictEqual(relationship.parent.id(), '7', 'house relationship parent found');
   });
 
-  test('relationshipReferenceFor hasMany', async function (assert) {
+  tes('relationshipReferenceFor hasMany', async function (assert) {
     assert.expect(3);
     this.owner.register('service:store', CustomStore);
     store = this.owner.lookup('service:store') as Store;
@@ -595,4 +597,5 @@ module('unit/model - Custom Class Model', function (hooks) {
     assert.strictEqual(relationship.type, 'house', 'house relationship type found');
     assert.strictEqual(relationship.parent.id(), '7', 'house relationship parent found');
   });
+  */
 });
