--- packages/store/addon/-private/system/references.ts
@@ -1,7 +1,0 @@
-/**
-  @module @ember-data/store
-*/
-
-import RecordReference from './references/record';
-
-export { RecordReference };
