+++ tsconfig.json
@@ -58,10 +58,6 @@
       "packages/store/addon/-private/system/snapshot-record-array.ts",
       "packages/store/addon/-private/system/schema-definition-service.ts",
       "packages/store/addon/-private/system/request-cache.ts",
-      "packages/store/addon/-private/system/references/belongs-to.ts",
-      "packages/store/addon/-private/system/references/has-many.ts",
-      "packages/store/addon/-private/system/references/reference.ts",
-      "packages/store/addon/-private/system/references/record.ts",
       "packages/store/addon/-private/system/record-notification-manager.ts",
       "packages/store/addon/-private/system/record-data-for.ts",
       "packages/store/addon/-private/system/record-arrays.ts",
